refactory the folder instruction
