--- XProject/src/Account/ChequeAccount.java
@@ -1,11 +1,0 @@
-package Account;
-
-/**
- * Created by jchen on 17/02/17.
- */
-public class ChequeAccount extends Account {
-    public ChequeAccount(String owner, long balance) {
-        super(owner, balance);
-    }
-
-}
