--- XProject/src/demo/practice/Polymorphism/Circle.java
@@ -1,16 +1,0 @@
-package demo.practice.Polymorphism;
-
-/**
- * Created by jchen on 19/02/17.
- */
-public class Circle implements IShape{
-    int x = 10;
-    public double area( ) {
-
-        return (6 * x * x);
-    }
-
-    public double volume() {
-        return (x * x * x);
-    }
-}
