--- XProject/src/demo/practice/Polymorphism/IShape.java
@@ -1,9 +1,0 @@
-package demo.practice.Polymorphism;
-
-/**
- * Created by jchen on 19/02/17.
- */
-public interface IShape {
-    public double area();
-    public double volume();
-}
