--- XProject/src/demo/practice/animal/IAnimal.java
@@ -1,18 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package demo.practice.animal;
-
-/**
- *
- * @author jotaiwan
- */
-public interface IAnimal {
-    public String Name = "Animal";	//animal should have name
-    
-    public String eat();            // animal eat
-    
-    public String sleep();          // animal sleep
-}
