--- XProject/src/demo/practice/animal/IBaby.java
@@ -1,14 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package demo.practice.animal;
-
-/**
- *
- * @author jotaiwan
- */
-public interface IBaby extends IAnimal, IPeople {
-    public boolean Cry();	//only baby love cry
-}
