--- XProject/src/demo/practice/animal/IPeople.java
@@ -1,18 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package demo.practice.animal;
-
-/**
- *
- * @author jotaiwan
- */
-public interface IPeople {
-    public int Age = 0;	
-    
-    public String Talk();	//only people can talk
-
-}
-
