--- XProject/src/demo/practice/messivealarm/MessiveDoor.java
@@ -1,10 +1,0 @@
-package demo.practice.messivealarm;
-
-/**
- * Created by jchen on 19/02/17.
- */
-public abstract class MessiveDoor {
-
-    abstract void open();
-    abstract void close();
-}
