+++ XProject/src/practice/Polymorphism/IShape.java
@@ -1,0 +1,9 @@
+package practice.Polymorphism;
+
+/**
+ * Created by jchen on 19/02/17.
+ */
+public interface IShape {
+    public double area();
+    public double volume();
+}
