+++ XProject/src/practice/alarmadvance/Door.java
@@ -1,0 +1,15 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package practice.alarmadvance;
+
+/**
+ *
+ * @author jotaiwan
+ */
+public abstract class Door {
+    public abstract void open();
+    public abstract void close();
+}
