+++ XProject/src/practice/classexample/SubClass.java
@@ -1,0 +1,24 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package practice.classexample;
+
+/**
+ *
+ * @author jotaiwan
+ */
+public class SubClass extends SuperClass {
+
+  public SubClass(){
+    super(300);
+    System.out.println("SuperClass");
+  }  
+  
+  public SubClass(int n){
+    System.out.println("SubClass(int n):"+n);
+    this.n = n;
+    }
+  
+}
