+++ XProject/src/practice/messivealarm/MessiveDoor.java
@@ -1,0 +1,10 @@
+package practice.messivealarm;
+
+/**
+ * Created by jchen on 19/02/17.
+ */
+public abstract class MessiveDoor {
+
+    abstract void open();
+    abstract void close();
+}
